02.ENH/smartthings-logic/requirements.txt
pandas>=2.0.0
python-dotenv>=1.0.0
langchain>=0.3.0
langchain-openai>=0.3.0
openpyxl>=3.1.5